+++ packages/dsl-visual-tests/jest.config.cjs
@@ -9,7 +9,6 @@ module.exports = {
   testMatch: ["<rootDir>/src/**/*.test.ts"],
   testPathIgnorePatterns: ["/node_modules/", "/dist/"],
   moduleNameMapper: {
-    "^@m0saic/platform/(.*)$": "<rootDir>/../platform/src/$1",
     "^@m0saic/dsl-file-formats$": "<rootDir>/../dsl-file-formats/src/index.ts",
     "^@m0saic/dsl$": "<rootDir>/../dsl/src/index.ts",
     "^@m0saic/dsl-stdlib$": "<rootDir>/../dsl-stdlib/src/index.ts",